open_wm: added code to fully reconnect if the connection is lost. vfd_driver: showint now allows multiple ints on the same display by only clearing the grids that are needed. This was used to display the temperature with two decimal places main: temperature now displayed with two decimals. No cases yet for temperatuers that do not fit the xx.xx format (9.00, 100.00, -10.00, etc)

+++ main/include/open_wm.h
@@ -34,10 +34,10 @@
 
 
 typedef struct owm_data_t {
-    int temp_now;
-    int temp_high;
-    int temp_low;
-    int temp_feel;
+    double temp_now;
+    double temp_high;
+    double temp_low;
+    double temp_feel;
 };
 
 static const char *TAG = "OWM REQUEST";
